Update trie code with refactored nibble paths
